+++ kinetic.js
@@ -7411,6 +7411,11 @@ var Kinetic = {};
                 child: child
             });
 
+            // if node under drag we need to update drag animation
+            if (child.isDragging()) {
+                Kinetic.DD.anim.setLayers(child.getLayer());
+            }
+
             // chainable
             return this;
         },